backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database drivers
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1
redis==5.0.1

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-generativeai==0.3.2

# Groq API
groq==0.4.1

# Slack
slack-sdk==3.26.2

# Twilio (SMS)
twilio==8.11.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
